refactored code and moved files to correct structure
