--- src/app/common/context/types.ts
@@ -1,7 +1,0 @@
-export interface AuthState {
-	isLoggedIn: boolean;
-	logOutUser: () => void;
-	setIsLoggedIn: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-export type AuthHeader = string | null;
