+++ src/app/common/hooks/useAuth.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { AuthContext } from '@/app/common/context/AuthContext';
-import { AuthState } from '@/app/common/context/types';
+import { AuthState } from '@/types/auth';
 
 const useAuth = (): AuthState => useContext(AuthContext);
 export default useAuth;
